+++ FractalProject/src/code/Fractals/Multibrot.java
@@ -7,12 +7,12 @@ public class Multibrot extends Fractal{
 	}
 	
 	public Multibrot(int rows, int cols){
-		super(rows, cols,-1f,1f,-1.3f,1.3f);
+		super(rows, cols,-1,1,-1.3,1.3);
 	}
 
 	@Override
-	public int calculate(float xCalc, float yCalc) {
-		float tX, tY, tmp;
+	public int calculate(double xCalc, double yCalc) {
+		double tX, tY, tmp;
 		double dist;
 		
 		// Mandelbrot starts every test from (0,0) AKA the origin
@@ -25,8 +25,8 @@ public class Multibrot extends Fractal{
 		dist = Math.sqrt(tX*tX + tY*tY); // Use math to find distance from 0
 		
 		while(dist < this.getEscapeDistance() && passes < this.getMaxEscapes()){
-			tmp = (float) (tX*tX*tX - (3.0*tX*tY*tY) + xCalc); // TEMP X: x' = x^2 - y^2 + (original x value)
-			tY = (float) (3.0 * tX*tX * tY - tY*tY*tY + yCalc); // y' = 2xy + (original y value)
+			tmp = tX*tX*tX - (3.0*tX*tY*tY) + xCalc; // TEMP X: x' = x^2 - y^2 + (original x value)
+			tY = 3.0 * tX*tX * tY - tY*tY*tY + yCalc; // y' = 2xy + (original y value)
 			tX = tmp; // set X to TEMP X
 			dist = Math.sqrt(tX*tX + tY*tY); // recalculate distance
 			passes++; // add a pass
